django_api/movies_admin/config/settings.py
import os

from dotenv import load_dotenv
from pathlib import Path
from split_settings.tools import include


load_dotenv()


# Build paths inside the project like this: BASE_DIR / 'subdir'.
BASE_DIR = Path(__file__).resolve().parent.parent

# SECURITY WARNING: keep the secret key used in production secret!
SECRET_KEY = os.environ.get('DJANGO_SECRET_KEY')

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = os.environ.get('DEBUG', False) == 'True'

ALLOWED_HOSTS = os.environ.get('ALLOWED_HOSTS').split(',')

CSRF_TRUSTED_ORIGINS = os.environ.get('CSRF_TRUSTED_ORIGINS').split(',')

include(
    'components/app_definition.py',
    'components/database.py',
    'components/validation.py',
    'components/internationalization.py',
    'components/static.py',
    'components/default.py'
)
