django_api/movies_admin/requirements.txt
Django==4.2.5
django-split-settings==1.2.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uwsgi==2.0.24